update mechanism of dividing people into teams

+++ src/main/java/agug/Person.java
@@ -3,9 +3,9 @@ package agug;
 class Person {
 
     String name;
-    int rate;
+    double rate;
 
-    public Person (String name, int rate) {
+    public Person (String name, double rate) {
         this.name = name;
         this.rate = rate;
     }
@@ -18,11 +18,11 @@ class Person {
         this.name = name;
     }
 
-    public int getRate() {
+    public double getRate() {
         return rate;
     }
 
-    public void setRate(int rate) {
+    public void setRate(double rate) {
         this.rate = rate;
     }
 }
